Refactor coq backend
